Create new ad integration manager. Move as stitch function from cue point manager to ne manager due to problems caused by stich doing impersonation when issuing API requests.
