+++ config/managers.ini.template
@@ -30,3 +30,7 @@ durationCoefficient = 1
 ;domain = @CUE_POINTS_SERVICE_HOSTNAME@
 ; domainPort = 80
 
+[adIntegration]
+;domain = @AD_INTEGRATION_SERVICE_HOSTNAME@
+; domainPort = 80
+
