+++ lib/managers/KalturaManifestManager.js
@@ -369,7 +369,7 @@ KalturaManifestManager.prototype.getRenditionFromCache = function(request, respo
 					playerConfig: playerConfig,
 					sessionId: sessionId
 			};
-			This.callPlayServerService('cuePoints', 'stitch', partnerId, stitchParams);
+			This.callPlayServerService('adIntegration', 'stitch', partnerId, stitchParams);
 		}
 	};
